Move Algorithms, Hash, PackageNevra class from data/models/package.py to type.py
