+++ alma_sbom/data/collectors/immudb/collector.py
@@ -2,7 +2,7 @@ import os
 from immudb_wrapper import ImmudbWrapper
 from logging import getLogger
 
-from alma_sbom.data import Package, PackageNevra, Hash, Algorithms
+from alma_sbom.data import Package, PackageNevra
 
 from .processor import DataProcessor, processor_factory
