+++ alma_sbom/type.py
@@ -1,4 +1,5 @@
 import argparse
+from dataclasses import dataclass
 from enum import Enum
 
 class SbomRecordType(Enum):
@@ -101,3 +102,89 @@ class SbomType:
     def __repr__(self) -> str:
         return f"{self.record_type.value}-{self.file_format_type.value}"
 
+### See the pythondx-python-libs Document: https://cyclonedx-python-library.readthedocs.io/en/latest/autoapi/cyclonedx/model/index.html#cyclonedx.model.HashAlgorithm
+class Algorithms(Enum):
+    SHA_256 = 'SHA-256'
+
+    @classmethod
+    def from_str(cls, string: str) -> 'Algorithms':
+        for alg in cls:
+            if string == alg.value:
+                return alg
+        raise ValueError(f'Invalid Algorithms string: {string}')
+
+@dataclass
+class Hash:
+    algorithm: Algorithms
+    value: str
+
+@dataclass
+class PackageNevra:
+    name: str = None
+    epoch: str = None
+    version: str = None
+    release: str = None
+    arch: str = None
+
+    def __repr__(self):
+        if self.epoch is not None:
+            return (
+                f'{self.epoch}:{self.name}-'
+                f'{self.version}-{self.release}.{self.arch}'
+            )
+        return f'{self.name}-{self.version}-' f'{self.release}.{self.arch}'
+
+    def get_NEVR(self) -> str:
+        if self.epoch is not None:
+            return (
+                f'{self.epoch}:{self.name}-{self.version}-{self.release}'
+            )
+        return f'{self.name}-{self.version}-{self.release}'
+
+    def get_EVR(self) -> str:
+        if self.epoch is not None:
+            return (
+                f'{self.epoch}:{self.version}-{self.release}'
+            )
+        return f'{self.version}-{self.release}'
+
+    ### TODO:
+    ## item 1
+    # need to be implemented out of PackageNevra or more outer ??
+    ## item 2
+    # need to be applied fix of normalize_epoch_in_cpe
+    # -> this need to be corvered by PackageNevra.epoch=='None'(str) ??
+    def get_cpe23(self) -> str:
+        cpe_version = '2.3'
+
+        cpe_epoch_part = f'{self.epoch if self.epoch else ""}'
+        cpe_epoch_part += '\\:' if cpe_epoch_part else ""
+        cpe = (
+            f'cpe:{cpe_version}:a:almalinux:'
+            f'{self.name}:{cpe_epoch_part}'
+            f'{self.version}-{self.release}:*:*:*:*:*:*:*'
+        )
+        return cpe
+
+    @classmethod
+    def from_str_has_epoch(package_name: str) -> 'PackageNevra':
+        raise NotImplementedError()
+
+    @classmethod
+    def from_str_nothas_epoch(cls, package_name: str) -> 'PackageNevra':
+        package_nevra = PackageNevra()
+
+        split_by_dot = package_name.replace('.rpm', '')[::-1].split('.', 1)
+        arch = split_by_dot[0][::-1]
+        split_by_hyphen = split_by_dot[1].split('-', 2)
+        release = split_by_hyphen[0][::-1]
+        version = split_by_hyphen[1][::-1]
+        name = split_by_hyphen[2][::-1]
+
+        return cls(
+            epoch = None,
+            name = name,
+            version = version,
+            release = release,
+            arch = arch,
+        )
